requirements.txt
certifi
python-ilorest-library
keyring
fabric
patchwork
pyyaml
natsort


